+++ lib/src/compatibility/http/http_response_handler.dart
@@ -187,5 +187,4 @@ class HttpResponseHandler {
       throw GenericError('Unexpected error: $e');
     }
   }
-
 }