+++ lib/utils/http_response_handler.dart
@@ -1,2 +1,0 @@
-export '../src/compatibility/http/http_response_handler.dart'
-    show HttpResponseHandler;